feat: Separate ScheduleContent
